[FE] create: 장소 선택 목록 페이지 분리
